refactor(utils): consolidate duplicate format functions into src/utils/ (#96)
Create src/utils/format.ts as single source of truth for formatCost,
formatTokens, and formatTimeAgo. Remove 7 duplicate definitions across
4 dashboard components. Update scan/shared.ts to re-export from
utils/format.ts for backward compatibility.
