+++ src/components/scan/shared.ts
@@ -1,17 +1,6 @@
 // CT Scan common utilities
-
-export const formatCost = (cost: number | undefined | null): string => {
-  if (cost == null || isNaN(cost) || cost <= 0) return '$0.00';
-  if (cost < 0.001) return `$${(cost * 1000).toFixed(2)}m`;
-  return `$${cost.toFixed(4)}`;
-};
-
-export const formatTokens = (n: number | undefined | null): string => {
-  if (n == null || isNaN(n)) return '0';
-  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}K`;
-  return String(n);
-};
+// Generic formatters re-exported from shared utils
+export { formatCost, formatTokens, formatTimeAgo } from '../../utils/format';
 
 export const getModelShort = (model: string): string => {
   if (model.includes('opus')) return 'Opus';
@@ -25,16 +14,6 @@ export const getModelColor = (model: string): string => {
   if (model.includes('sonnet')) return '#3b82f6';
   if (model.includes('haiku')) return '#10b981';
   return '#6b7280';
-};
-
-export const formatTimeAgo = (ts: string): string => {
-  const diff = Date.now() - new Date(ts).getTime();
-  const mins = Math.floor(diff / 60000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const hrs = Math.floor(mins / 60);
-  if (hrs < 24) return `${hrs}h ago`;
-  return `${Math.floor(hrs / 24)}d ago`;
 };
 
 export const CATEGORY_COLORS: Record<string, string> = {
